Dev to master

+++ src/project/api/booking.py
@@ -1,13 +1,19 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
+from project.schemas.user import UserSchema
 from project.schemas.booking import BookingSchema, BookingCreateUpdateSchema
 from project.core.exceptions.booking import BookingNotFound, BookingBadForeignKey, BookingBadDate
-from project.api.depends import database, booking_repo
+from project.api.depends import database, booking_repo, get_current_user, check_for_admin_access
 
 router = APIRouter()
 
 
-@router.get("/all_bookings", response_model=list[BookingSchema], status_code=status.HTTP_200_OK)
+@router.get(
+    "/all_bookings",
+    response_model=list[BookingSchema],
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
+)
 async def get_all_bookings() -> list[BookingSchema]:
     async with database.session() as session:
         all_bookings = await booking_repo.get_all_bookings(session=session)
@@ -15,7 +21,12 @@ async def get_all_bookings() -> list[BookingSchema]:
     return all_bookings
 
 
-@router.get("/booking/{booking_id}", response_model=BookingSchema, status_code=status.HTTP_200_OK)
+@router.get(
+    "/booking/{booking_id}",
+    response_model=BookingSchema,
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
+)
 async def get_booking_by_id(
     booking_id: int,
 ) -> BookingSchema:
@@ -27,7 +38,12 @@ async def get_booking_by_id(
     return booking
 
 
-@router.post("/add_booking", response_model=BookingSchema, status_code=status.HTTP_201_CREATED)
+@router.post(
+    "/add_booking",
+    response_model=BookingSchema,
+    status_code=status.HTTP_201_CREATED,
+    dependencies=[Depends(get_current_user)],
+)
 async def add_booking(
     booking_dto: BookingCreateUpdateSchema,
 ) -> BookingSchema:
@@ -43,6 +59,7 @@ async def add_booking(
     "/update_booking/{booking_id}",
     response_model=BookingSchema,
     status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
 )
 async def update_booking(
     booking_id: int,
@@ -63,7 +80,9 @@ async def update_booking(
 @router.delete("/delete_booking/{booking_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_booking(
     booking_id: int,
+    current_user: UserSchema = Depends(get_current_user),
 ) -> None:
+    check_for_admin_access(current_user)
     try:
         async with database.session() as session:
             booking = await booking_repo.delete_booking(session=session, booking_id=booking_id)
