+++ src/project/api/buyer.py
@@ -1,13 +1,19 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
+from project.schemas.user import UserSchema
 from project.schemas.buyer import BuyerSchema, BuyerCreateUpdateSchema
 from project.core.exceptions.buyer import BuyerNotFound
-from project.api.depends import database, buyer_repo
+from project.api.depends import database, buyer_repo, get_current_user, check_for_admin_access
 
 router = APIRouter()
 
 
-@router.get("/all_buyers", response_model=list[BuyerSchema], status_code=status.HTTP_200_OK)
+@router.get(
+    "/all_buyers",
+    response_model=list[BuyerSchema],
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
+)
 async def get_all_buyers() -> list[BuyerSchema]:
     async with database.session() as session:
         all_buyers = await buyer_repo.get_all_buyers(session=session)
@@ -15,7 +21,12 @@ async def get_all_buyers() -> list[BuyerSchema]:
     return all_buyers
 
 
-@router.get("/buyer/{buyer_id}", response_model=BuyerSchema, status_code=status.HTTP_200_OK)
+@router.get(
+    "/buyer/{buyer_id}",
+    response_model=BuyerSchema,
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
+)
 async def get_buyer_by_id(
     buyer_id: int,
 ) -> BuyerSchema:
@@ -27,7 +38,12 @@ async def get_buyer_by_id(
     return buyer
 
 
-@router.post("/add_buyer", response_model=BuyerSchema, status_code=status.HTTP_201_CREATED)
+@router.post(
+    "/add_buyer",
+    response_model=BuyerSchema,
+    status_code=status.HTTP_201_CREATED,
+    dependencies=[Depends(get_current_user)],
+)
 async def add_buyer(
     buyer_dto: BuyerCreateUpdateSchema,
 ) -> BuyerSchema:
@@ -40,6 +56,7 @@ async def add_buyer(
     "/update_buyer/{buyer_id}",
     response_model=BuyerSchema,
     status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
 )
 async def update_buyer(
     buyer_id: int,
@@ -60,7 +77,9 @@ async def update_buyer(
 @router.delete("/delete_buyer/{buyer_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_buyer(
     buyer_id: int,
+    current_user: UserSchema = Depends(get_current_user),
 ) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             buyer = await buyer_repo.delete_buyer(session=session, buyer_id=buyer_id)
